ToolsConfig tests
