+++ tests/groovy/ToolsConfig/MavenTest.groovy
@@ -4,19 +4,6 @@ handler = 'Tools'
 configHandler = evaluate(new File("/usr/share/jenkins/config-handlers/${handler}Config.groovy"))
 
 
-def testOracleUserPassword(){
-	def config = new Yaml().load("""
-oracle_jdk_download:
-  username: oracle-user
-  password: oracle-password
-""")
-
-	configHandler.setup(config)
-	def jdkInstaller = jenkins.model.Jenkins.instance.getDescriptor(hudson.tools.JDKInstaller)
-	assert jdkInstaller.username == 'oracle-user'
-	assert jdkInstaller.password.toString() == 'oracle-password'
-}
-
 def testMavenToolAutoInstaller(){
 	def config = new Yaml().load("""
 installations:
@@ -45,6 +32,5 @@ installations:
 	assert installation.home == '/user/share/maven-3.5.1'
 }
 
-testOracleUserPassword()
 testMavenToolAutoInstaller()
 testMavenToolManual()
